Move stuff into base subdir
