+++ src/api/__main__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 from api import app
-import config
+import base.config as config
 
 
 def main():
